+++ modules/Movecraft/src/main/java/net/countercraft/movecraft/listener/BlockBreakListener.java
@@ -1,0 +1,47 @@
+package net.countercraft.movecraft.listener;
+
+import net.countercraft.movecraft.MovecraftLocation;
+import net.countercraft.movecraft.config.Settings;
+import net.countercraft.movecraft.craft.Craft;
+import net.countercraft.movecraft.craft.CraftManager;
+import net.countercraft.movecraft.localisation.I18nSupport;
+import net.countercraft.movecraft.utils.MathUtils;
+import org.bukkit.ChatColor;
+import org.bukkit.Material;
+import org.bukkit.block.Sign;
+import org.bukkit.event.EventHandler;
+import org.bukkit.event.EventPriority;
+import org.bukkit.event.Listener;
+import org.bukkit.event.block.BlockBreakEvent;
+
+public class BlockBreakListener implements Listener {
+    @EventHandler(priority = EventPriority.HIGHEST)
+    public void onBlockBreak(final BlockBreakEvent e) {
+        if (e.isCancelled()) {
+            return;
+        }
+        if (e.getBlock().getType() == Material.WALL_SIGN) {
+            Sign s = (Sign) e.getBlock().getState();
+            if (s.getLine(0).equalsIgnoreCase(ChatColor.RED + I18nSupport.getInternationalisedString("Region Damaged"))) {
+                e.setCancelled(true);
+                return;
+            }
+        }
+        if (Settings.ProtectPilotedCrafts) {
+            MovecraftLocation mloc = MathUtils.bukkit2MovecraftLoc(e.getBlock().getLocation());
+            CraftManager.getInstance().getCraftsInWorld(e.getBlock().getWorld());
+            for (Craft craft : CraftManager.getInstance().getCraftsInWorld(e.getBlock().getWorld())) {
+                if (craft == null || craft.getDisabled()) {
+                    continue;
+                }
+                for (MovecraftLocation tloc : craft.getHitBox()) {
+                    if (tloc.equals(mloc)) {
+                        e.getPlayer().sendMessage(I18nSupport.getInternationalisedString("Player - Block part of piloted craft"));
+                        e.setCancelled(true);
+                        return;
+                    }
+                }
+            }
+        }
+    }
+}
