+++ modules/Movecraft/src/main/java/net/countercraft/movecraft/listener/BlockDispenseListener.java
@@ -1,0 +1,24 @@
+package net.countercraft.movecraft.listener;
+
+import net.countercraft.movecraft.craft.Craft;
+import net.countercraft.movecraft.craft.CraftManager;
+import net.countercraft.movecraft.utils.MathUtils;
+import org.bukkit.event.EventHandler;
+import org.bukkit.event.EventPriority;
+import org.bukkit.event.Listener;
+import org.bukkit.event.block.BlockDispenseEvent;
+
+public class BlockDispenseListener implements Listener {
+    @EventHandler(priority = EventPriority.HIGHEST)
+    public void onBlockDispense(BlockDispenseEvent e) {
+        CraftManager.getInstance().getCraftsInWorld(e.getBlock().getWorld());
+        for (Craft craft : CraftManager.getInstance().getCraftsInWorld(e.getBlock().getWorld())) {
+            if (craft != null &&
+                    !craft.isNotProcessing() &&
+                    MathUtils.locIsNearCraftFast(craft, MathUtils.bukkit2MovecraftLoc(e.getBlock().getLocation()))) {
+                e.setCancelled(true);
+                return;
+            }
+        }
+    }
+}
