+++ modules/Movecraft/src/main/java/net/countercraft/movecraft/listener/BlockFromToListener.java
@@ -1,0 +1,51 @@
+package net.countercraft.movecraft.listener;
+
+import net.countercraft.movecraft.MovecraftLocation;
+import net.countercraft.movecraft.config.Settings;
+import net.countercraft.movecraft.craft.Craft;
+import net.countercraft.movecraft.craft.CraftManager;
+import net.countercraft.movecraft.utils.MathUtils;
+import org.bukkit.Material;
+import org.bukkit.block.Block;
+import org.bukkit.event.EventHandler;
+import org.bukkit.event.EventPriority;
+import org.bukkit.event.Listener;
+import org.bukkit.event.block.BlockFromToEvent;
+
+public class BlockFromToListener implements Listener {
+    // TODO: Check on this double logic
+
+    // prevent water and lava from spreading on moving crafts
+    @EventHandler(priority = EventPriority.HIGHEST)
+    public void onBlockFromTo(BlockFromToEvent e) {
+        if (e.isCancelled()) {
+            return;
+        }
+        Block block = e.getToBlock();
+        if (block.getType() != Material.WATER && block.getType() != Material.LAVA) {
+            return;
+        }
+        for (Craft tcraft : CraftManager.getInstance().getCraftsInWorld(block.getWorld())) {
+            if ((!tcraft.isNotProcessing()) && MathUtils.locIsNearCraftFast(tcraft, MathUtils.bukkit2MovecraftLoc(block.getLocation()))) {
+                e.setCancelled(true);
+                return;
+            }
+        }
+    }
+
+    @EventHandler
+    public void onFlow(BlockFromToEvent e){
+        if(Settings.DisableSpillProtection)
+            return;
+        if(!e.getBlock().isLiquid())
+            return;
+        MovecraftLocation loc = MathUtils.bukkit2MovecraftLoc(e.getBlock().getLocation());
+        MovecraftLocation toLoc = MathUtils.bukkit2MovecraftLoc(e.getToBlock().getLocation());
+        for(Craft craft : CraftManager.getInstance().getCraftsInWorld(e.getBlock().getWorld())){
+            if(craft.getHitBox().contains((loc)) && !craft.getFluidLocations().contains(toLoc)) {
+                e.setCancelled(true);
+                break;
+            }
+        }
+    }
+}
