+++ modules/Movecraft/src/main/java/net/countercraft/movecraft/listener/BlockIgniteListener.java
@@ -1,0 +1,64 @@
+package net.countercraft.movecraft.listener;
+
+import com.sk89q.worldguard.protection.ApplicableRegionSet;
+import com.sk89q.worldguard.protection.flags.DefaultFlag;
+import net.countercraft.movecraft.Movecraft;
+import net.countercraft.movecraft.config.Settings;
+import net.countercraft.movecraft.craft.Craft;
+import net.countercraft.movecraft.craft.CraftManager;
+import net.countercraft.movecraft.utils.MathUtils;
+import org.bukkit.Location;
+import org.bukkit.block.Block;
+import org.bukkit.event.EventHandler;
+import org.bukkit.event.EventPriority;
+import org.bukkit.event.Listener;
+import org.bukkit.event.block.BlockIgniteEvent;
+import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
+
+public class BlockIgniteListener implements Listener {
+    @EventHandler(priority = EventPriority.LOW)
+    public void onBlockIgnite(BlockIgniteEvent event) {
+        if (event.isCancelled()) {
+            return;
+        }
+        final Craft adjacentCraft = adjacentCraft(event.getBlock().getLocation());
+        // replace blocks with fire occasionally, to prevent fast craft from simply ignoring fire
+        if (Settings.FireballPenetration && event.getCause() == BlockIgniteEvent.IgniteCause.FIREBALL) {
+            Block testBlock = event.getBlock().getRelative(-1, 0, 0);
+            if (!testBlock.getType().isBurnable())
+                testBlock = event.getBlock().getRelative(1, 0, 0);
+            if (!testBlock.getType().isBurnable())
+                testBlock = event.getBlock().getRelative(0, 0, -1);
+            if (!testBlock.getType().isBurnable())
+                testBlock = event.getBlock().getRelative(0, 0, 1);
+
+            if (!testBlock.getType().isBurnable()) {
+                return;
+            }
+            // check to see if fire spread is allowed, don't check if worldguard integration is not enabled
+            if (Movecraft.getInstance().getWorldGuardPlugin() != null && (Settings.WorldGuardBlockMoveOnBuildPerm || Settings.WorldGuardBlockSinkOnPVPPerm)) {
+                ApplicableRegionSet set = Movecraft.getInstance().getWorldGuardPlugin().getRegionManager(testBlock.getWorld()).getApplicableRegions(testBlock.getLocation());
+                if (!set.allows(DefaultFlag.FIRE_SPREAD)) {
+                    return;
+                }
+            }
+            testBlock.setType(org.bukkit.Material.AIR);
+        }
+        else if (adjacentCraft != null) {
+            adjacentCraft.getHitBox().add(MathUtils.bukkit2MovecraftLoc(event.getBlock().getLocation()));
+        }
+
+    }
+
+    @Nullable
+    private Craft adjacentCraft(@NotNull Location location) {
+        for (Craft craft : CraftManager.getInstance().getCraftsInWorld(location.getWorld())) {
+            if (!MathUtils.locIsNearCraftFast(craft, MathUtils.bukkit2MovecraftLoc(location))) {
+                continue;
+            }
+            return craft;
+        }
+        return null;
+    }
+}
