+++ modules/Movecraft/src/main/java/net/countercraft/movecraft/listener/BlockListener.java
@@ -19,7 +19,6 @@ package net.countercraft.movecraft.listener;
 
 import com.sk89q.worldguard.bukkit.WorldGuardPlugin;
 import com.sk89q.worldguard.protection.ApplicableRegionSet;
-import com.sk89q.worldguard.protection.flags.DefaultFlag;
 import net.countercraft.movecraft.Movecraft;
 import net.countercraft.movecraft.MovecraftLocation;
 import net.countercraft.movecraft.config.Settings;
@@ -32,9 +31,6 @@ import org.bukkit.*;
 import org.bukkit.block.Block;
 import org.bukkit.block.BlockFace;
 import org.bukkit.block.Hopper;
-import org.bukkit.block.Sign;
-import org.bukkit.entity.Entity;
-import org.bukkit.entity.EntityType;
 import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.EventPriority;
@@ -43,18 +39,11 @@ import org.bukkit.event.block.*;
 import org.bukkit.event.entity.EntityExplodeEvent;
 import org.bukkit.event.entity.ItemSpawnEvent;
 import org.bukkit.event.inventory.InventoryMoveItemEvent;
-import org.bukkit.material.Attachable;
-import org.bukkit.material.MaterialData;
-import org.bukkit.material.PistonBaseMaterial;
 import org.bukkit.scheduler.BukkitRunnable;
-import org.bukkit.util.Vector;
-import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
 
 import java.util.Arrays;
 import java.util.Iterator;
 import java.util.List;
-import java.util.Random;
 
 public class BlockListener implements Listener {
 
@@ -87,36 +76,6 @@ public class BlockListener implements Listener {
         }
     }
 
-    @EventHandler(priority = EventPriority.HIGHEST)
-    public void onBlockBreak(final BlockBreakEvent e) {
-        if (e.isCancelled()) {
-            return;
-        }
-        if (e.getBlock().getType() == Material.WALL_SIGN) {
-            Sign s = (Sign) e.getBlock().getState();
-            if (s.getLine(0).equalsIgnoreCase(ChatColor.RED + I18nSupport.getInternationalisedString("Region Damaged"))) {
-                e.setCancelled(true);
-                return;
-            }
-        }
-        if (Settings.ProtectPilotedCrafts) {
-            MovecraftLocation mloc = MathUtils.bukkit2MovecraftLoc(e.getBlock().getLocation());
-            CraftManager.getInstance().getCraftsInWorld(e.getBlock().getWorld());
-            for (Craft craft : CraftManager.getInstance().getCraftsInWorld(e.getBlock().getWorld())) {
-                if (craft == null || craft.getDisabled()) {
-                    continue;
-                }
-                for (MovecraftLocation tloc : craft.getHitBox()) {
-                    if (tloc.equals(mloc)) {
-                        e.getPlayer().sendMessage(I18nSupport.getInternationalisedString("Player - Block part of piloted craft"));
-                        e.setCancelled(true);
-                        return;
-                    }
-                }
-            }
-        }
-    }
-
     // prevent items from dropping from moving crafts
     @EventHandler(priority = EventPriority.HIGHEST)
     public void onItemSpawn(final ItemSpawnEvent e) {
@@ -126,55 +85,6 @@ public class BlockListener implements Listener {
         for (Craft tcraft : CraftManager.getInstance().getCraftsInWorld(e.getLocation().getWorld())) {
             if ((!tcraft.isNotProcessing()) && MathUtils.locationInHitBox(tcraft.getHitBox(), e.getLocation())) {
                 e.setCancelled(true);
-                return;
-            }
-        }
-    }
-
-    // prevent water and lava from spreading on moving crafts
-    @EventHandler(priority = EventPriority.HIGHEST)
-    public void onBlockFromTo(BlockFromToEvent e) {
-        if (e.isCancelled()) {
-            return;
-        }
-        Block block = e.getToBlock();
-        if (block.getType() != Material.WATER && block.getType() != Material.LAVA) {
-            return;
-        }
-        for (Craft tcraft : CraftManager.getInstance().getCraftsInWorld(block.getWorld())) {
-            if ((!tcraft.isNotProcessing()) && MathUtils.locIsNearCraftFast(tcraft, MathUtils.bukkit2MovecraftLoc(block.getLocation()))) {
-                e.setCancelled(true);
-                return;
-            }
-        }
-    }
-
-    // process certain redstone on cruising crafts
-    @EventHandler(priority = EventPriority.HIGHEST)
-    public void onRedstoneEvent(BlockRedstoneEvent event) {
-        Block block = event.getBlock();
-        CraftManager.getInstance().getCraftsInWorld(block.getWorld());
-        for (Craft tcraft : CraftManager.getInstance().getCraftsInWorld(block.getWorld())) {
-            MovecraftLocation mloc = new MovecraftLocation(block.getX(), block.getY(), block.getZ());
-            if (MathUtils.locIsNearCraftFast(tcraft, mloc) &&
-                    tcraft.getCruising() && (block.getTypeId() == 29 ||
-                    block.getTypeId() == 33 || block.getTypeId() == 23 &&
-                    !tcraft.isNotProcessing())) {
-                event.setNewCurrent(event.getOldCurrent()); // don't allow piston movement on cruising crafts
-                return;
-            }
-        }
-    }
-
-    // prevent pistons on cruising crafts
-    @EventHandler(priority = EventPriority.HIGHEST)
-    public void onPistonEvent(BlockPistonExtendEvent event) {
-        Block block = event.getBlock();
-        CraftManager.getInstance().getCraftsInWorld(block.getWorld());
-        for (Craft tcraft : CraftManager.getInstance().getCraftsInWorld(block.getWorld())) {
-            MovecraftLocation mloc = new MovecraftLocation(block.getX(), block.getY(), block.getZ());
-            if (MathUtils.locIsNearCraftFast(tcraft, mloc) && tcraft.getCruising() && !tcraft.isNotProcessing()) {
-                event.setCancelled(true);
                 return;
             }
         }
@@ -193,103 +103,6 @@ public class BlockListener implements Listener {
             if (MathUtils.locIsNearCraftFast(tcraft, mloc) && tcraft.getCruising() && !tcraft.isNotProcessing()) {
                 event.setCancelled(true);
                 return;
-            }
-        }
-    }
-
-    // prevent fragile items from dropping on cruising crafts
-    @EventHandler(priority = EventPriority.HIGHEST)
-    public void onPhysics(BlockPhysicsEvent event) {
-        if (event.isCancelled()) {
-            return;
-        }
-
-        Block block = event.getBlock();
-
-        final int[] fragileBlocks = new int[]{26, 34, 50, 55, 63, 64, 65, 68, 69, 70, 71, 72, 75, 76, 77, 93, 94, 96, 131, 132, 143, 147, 148, 149, 150, 151, 171, 193, 194, 195, 196, 197};
-        CraftManager.getInstance().getCraftsInWorld(block.getWorld());
-        for (Craft tcraft : CraftManager.getInstance().getCraftsInWorld(block.getWorld())) {
-            MovecraftLocation mloc = new MovecraftLocation(block.getX(), block.getY(), block.getZ());
-            if (!MathUtils.locIsNearCraftFast(tcraft, mloc)) {
-                continue;
-            }
-            if (Arrays.binarySearch(fragileBlocks, block.getTypeId()) >= 0) {
-                MaterialData m = block.getState().getData();
-                BlockFace face = BlockFace.DOWN;
-                boolean faceAlwaysDown = false;
-                if (block.getTypeId() == 149 || block.getTypeId() == 150 || block.getTypeId() == 93 || block.getTypeId() == 94)
-                    faceAlwaysDown = true;
-                if (m instanceof Attachable && !faceAlwaysDown) {
-                    face = ((Attachable) m).getAttachedFace();
-                }
-                if (!event.getBlock().getRelative(face).getType().isSolid()) {
-//						if(event.getEventName().equals("BlockPhysicsEvent")) {
-                    event.setCancelled(true);
-                    return;
-                }
-            }
-        }
-    }
-
-    @EventHandler(priority = EventPriority.LOW)
-    public void onBlockIgnite(BlockIgniteEvent event) {
-        if (event.isCancelled()) {
-            return;
-        }
-        final Craft adjacentCraft = adjacentCraft(event.getBlock().getLocation());
-        // replace blocks with fire occasionally, to prevent fast craft from simply ignoring fire
-        if (Settings.FireballPenetration && event.getCause() == BlockIgniteEvent.IgniteCause.FIREBALL) {
-            Block testBlock = event.getBlock().getRelative(-1, 0, 0);
-            if (!testBlock.getType().isBurnable())
-                testBlock = event.getBlock().getRelative(1, 0, 0);
-            if (!testBlock.getType().isBurnable())
-                testBlock = event.getBlock().getRelative(0, 0, -1);
-            if (!testBlock.getType().isBurnable())
-                testBlock = event.getBlock().getRelative(0, 0, 1);
-
-            if (!testBlock.getType().isBurnable()) {
-                return;
-            }
-            // check to see if fire spread is allowed, don't check if worldguard integration is not enabled
-            if (Movecraft.getInstance().getWorldGuardPlugin() != null && (Settings.WorldGuardBlockMoveOnBuildPerm || Settings.WorldGuardBlockSinkOnPVPPerm)) {
-                ApplicableRegionSet set = Movecraft.getInstance().getWorldGuardPlugin().getRegionManager(testBlock.getWorld()).getApplicableRegions(testBlock.getLocation());
-                if (!set.allows(DefaultFlag.FIRE_SPREAD)) {
-                    return;
-                }
-            }
-            testBlock.setType(org.bukkit.Material.AIR);
-        } else if (adjacentCraft != null) {
-
-            adjacentCraft.getHitBox().add(MathUtils.bukkit2MovecraftLoc(event.getBlock().getLocation()));
-        }
-
-    }
-
-    @EventHandler(priority = EventPriority.HIGHEST)
-    public void onBlockDispense(BlockDispenseEvent e) {
-        CraftManager.getInstance().getCraftsInWorld(e.getBlock().getWorld());
-        for (Craft craft : CraftManager.getInstance().getCraftsInWorld(e.getBlock().getWorld())) {
-            if (craft != null &&
-                    !craft.isNotProcessing() &&
-                    MathUtils.locIsNearCraftFast(craft, MathUtils.bukkit2MovecraftLoc(e.getBlock().getLocation()))) {
-                e.setCancelled(true);
-                return;
-            }
-        }
-    }
-
-    @EventHandler
-    public void onFlow(BlockFromToEvent e){
-        if(Settings.DisableSpillProtection)
-            return;
-        if(!e.getBlock().isLiquid())
-            return;
-        MovecraftLocation loc = MathUtils.bukkit2MovecraftLoc(e.getBlock().getLocation());
-        MovecraftLocation toLoc = MathUtils.bukkit2MovecraftLoc(e.getToBlock().getLocation());
-        for(Craft craft : CraftManager.getInstance().getCraftsInWorld(e.getBlock().getWorld())){
-            if(craft.getHitBox().contains((loc)) && !craft.getFluidLocations().contains(toLoc)) {
-                e.setCancelled(true);
-                break;
             }
         }
     }
@@ -367,34 +180,5 @@ public class BlockListener implements Listener {
             }
         }
 
-    }
-
-    @Nullable
-    private Craft adjacentCraft(@NotNull Location location) {
-        for (Craft craft : CraftManager.getInstance().getCraftsInWorld(location.getWorld())) {
-            if (!MathUtils.locIsNearCraftFast(craft, MathUtils.bukkit2MovecraftLoc(location))) {
-                continue;
-            }
-            return craft;
-        }
-        return null;
-    }
-
-    private boolean pistonFacingLocation(Location loc) {
-        final Vector[] SHIFTS = {new Vector(0,1,0), new Vector(0,-1,0),
-                new Vector(1,0,0), new Vector(-1,0,0),
-                new Vector(0,0,1), new Vector(0,0,-1)};
-        for (Vector shift : SHIFTS) {
-            final Location test = loc.add(shift);
-            if (!(test.getBlock().getState().getData() instanceof PistonBaseMaterial)) {
-                continue;
-            }
-            PistonBaseMaterial piston = (PistonBaseMaterial) test.getBlock().getState().getData();
-            if (!test.getBlock().getRelative(piston.getFacing()).getLocation().equals(loc)) {
-                continue;
-            }
-            return true;
-        }
-        return false;
     }
 }
