+++ modules/Movecraft/src/main/java/net/countercraft/movecraft/listener/BlockPhysicsListener.java
@@ -1,0 +1,52 @@
+package net.countercraft.movecraft.listener;
+
+import net.countercraft.movecraft.MovecraftLocation;
+import net.countercraft.movecraft.craft.Craft;
+import net.countercraft.movecraft.craft.CraftManager;
+import net.countercraft.movecraft.utils.MathUtils;
+import org.bukkit.block.Block;
+import org.bukkit.block.BlockFace;
+import org.bukkit.event.EventHandler;
+import org.bukkit.event.EventPriority;
+import org.bukkit.event.Listener;
+import org.bukkit.event.block.BlockPhysicsEvent;
+import org.bukkit.material.Attachable;
+import org.bukkit.material.MaterialData;
+
+import java.util.Arrays;
+
+public class BlockPhysicsListener implements Listener {
+    // prevent fragile items from dropping on cruising crafts
+    @EventHandler(priority = EventPriority.HIGHEST)
+    public void onPhysics(BlockPhysicsEvent event) {
+        if (event.isCancelled()) {
+            return;
+        }
+
+        Block block = event.getBlock();
+
+        final int[] fragileBlocks = new int[]{26, 34, 50, 55, 63, 64, 65, 68, 69, 70, 71, 72, 75, 76, 77, 93, 94, 96, 131, 132, 143, 147, 148, 149, 150, 151, 171, 193, 194, 195, 196, 197};
+        CraftManager.getInstance().getCraftsInWorld(block.getWorld());
+        for (Craft tcraft : CraftManager.getInstance().getCraftsInWorld(block.getWorld())) {
+            MovecraftLocation mloc = new MovecraftLocation(block.getX(), block.getY(), block.getZ());
+            if (!MathUtils.locIsNearCraftFast(tcraft, mloc)) {
+                continue;
+            }
+            if (Arrays.binarySearch(fragileBlocks, block.getTypeId()) >= 0) {
+                MaterialData m = block.getState().getData();
+                BlockFace face = BlockFace.DOWN;
+                boolean faceAlwaysDown = false;
+                if (block.getTypeId() == 149 || block.getTypeId() == 150 || block.getTypeId() == 93 || block.getTypeId() == 94)
+                    faceAlwaysDown = true;
+                if (m instanceof Attachable && !faceAlwaysDown) {
+                    face = ((Attachable) m).getAttachedFace();
+                }
+                if (!event.getBlock().getRelative(face).getType().isSolid()) {
+//						if(event.getEventName().equals("BlockPhysicsEvent")) {
+                    event.setCancelled(true);
+                    return;
+                }
+            }
+        }
+    }
+}
