+++ modules/Movecraft/src/main/java/net/countercraft/movecraft/listener/BlockPistonListener.java
@@ -1,0 +1,27 @@
+package net.countercraft.movecraft.listener;
+
+import net.countercraft.movecraft.MovecraftLocation;
+import net.countercraft.movecraft.craft.Craft;
+import net.countercraft.movecraft.craft.CraftManager;
+import net.countercraft.movecraft.utils.MathUtils;
+import org.bukkit.block.Block;
+import org.bukkit.event.EventHandler;
+import org.bukkit.event.EventPriority;
+import org.bukkit.event.Listener;
+import org.bukkit.event.block.BlockPistonExtendEvent;
+
+public class BlockPistonListener implements Listener {
+    // prevent pistons on cruising crafts
+    @EventHandler(priority = EventPriority.HIGHEST)
+    public void onPistonEvent(BlockPistonExtendEvent event) {
+        Block block = event.getBlock();
+        CraftManager.getInstance().getCraftsInWorld(block.getWorld());
+        for (Craft tcraft : CraftManager.getInstance().getCraftsInWorld(block.getWorld())) {
+            MovecraftLocation mloc = new MovecraftLocation(block.getX(), block.getY(), block.getZ());
+            if (MathUtils.locIsNearCraftFast(tcraft, mloc) && tcraft.getCruising() && !tcraft.isNotProcessing()) {
+                event.setCancelled(true);
+                return;
+            }
+        }
+    }
+}
