+++ modules/Movecraft/src/main/java/net/countercraft/movecraft/listener/BlockRedstoneListener.java
@@ -1,0 +1,30 @@
+package net.countercraft.movecraft.listener;
+
+import net.countercraft.movecraft.MovecraftLocation;
+import net.countercraft.movecraft.craft.Craft;
+import net.countercraft.movecraft.craft.CraftManager;
+import net.countercraft.movecraft.utils.MathUtils;
+import org.bukkit.block.Block;
+import org.bukkit.event.EventHandler;
+import org.bukkit.event.EventPriority;
+import org.bukkit.event.Listener;
+import org.bukkit.event.block.BlockRedstoneEvent;
+
+public class BlockRedstoneListener implements Listener {
+    // process certain redstone on cruising crafts
+    @EventHandler(priority = EventPriority.HIGHEST)
+    public void onRedstoneEvent(BlockRedstoneEvent event) {
+        Block block = event.getBlock();
+        CraftManager.getInstance().getCraftsInWorld(block.getWorld());
+        MovecraftLocation mloc = new MovecraftLocation(block.getX(), block.getY(), block.getZ());
+        for (Craft tcraft : CraftManager.getInstance().getCraftsInWorld(block.getWorld())) {
+            if (MathUtils.locIsNearCraftFast(tcraft, mloc) &&
+                    tcraft.getCruising() && (block.getTypeId() == 29 ||
+                    block.getTypeId() == 33 || block.getTypeId() == 23 &&
+                    !tcraft.isNotProcessing())) {
+                event.setNewCurrent(event.getOldCurrent()); // don't allow piston movement on cruising crafts
+                return;
+            }
+        }
+    }
+}
